Move common functions out
